+++ packages/graphql/src/actions/BaseActions.ts
@@ -52,5 +52,6 @@ export abstract class BaseActions {
 
   abstract createProjectBase(input: NxsMutationArgs<'addProject'>['input']): ProjectBaseContract | Promise<ProjectBaseContract>
   abstract authenticate (): Promise<void>
+  abstract logout (): Promise<void>
   abstract getRuns ({ projectId }: { projectId: string }): Promise<void>
 }
