+++ packages/graphql/src/context/BaseContext.ts
@@ -12,7 +12,7 @@ import type { Project } from '../entities/Project'
 export abstract class BaseContext {
   abstract readonly actions: BaseActions
   abstract projects: Project[]
-  abstract user?: User
+  abstract user?: User = undefined
 
   wizard = new Wizard()
   app = new App(this)
