+++ packages/graphql/src/gen/nxs.gen.ts
@@ -6,15 +6,15 @@
 
 
 import type { BaseContext } from "./../context/BaseContext"
-import type { App } from "./../entities/App.js"
-import type { Project } from "./../entities/Project.js"
-import type { Query } from "./../entities/Query.js"
-import type { User } from "./../entities/User.js"
-import type { TestingTypeInfo } from "./../entities/TestingTypeInfo.js"
-import type { Wizard } from "./../entities/Wizard.js"
-import type { WizardBundler } from "./../entities/WizardBundler.js"
-import type { WizardFrontendFramework } from "./../entities/WizardFrontendFramework.js"
-import type { WizardNpmPackage } from "./../entities/WizardNpmPackage.js"
+import type { App } from "./../entities/App"
+import type { Project } from "./../entities/Project"
+import type { Query } from "./../entities/Query"
+import type { User } from "./../entities/User"
+import type { TestingTypeInfo } from "./../entities/TestingTypeInfo"
+import type { Wizard } from "./../entities/Wizard"
+import type { WizardBundler } from "./../entities/WizardBundler"
+import type { WizardFrontendFramework } from "./../entities/WizardFrontendFramework"
+import type { WizardNpmPackage } from "./../entities/WizardNpmPackage"
 import type { core } from "nexus"
 declare global {
   interface NexusGenCustomInputMethods<TypeName extends string> {
@@ -109,6 +109,7 @@ export interface NexusGenFieldTypes {
     appCreateConfigFile: NexusGenRootTypes['App'] | null; // App
     authenticate: NexusGenRootTypes['App'] | null; // App
     initializePlugins: NexusGenRootTypes['Project'] | null; // Project
+    logout: NexusGenRootTypes['App'] | null; // App
     wizardInstallDependencies: NexusGenRootTypes['Wizard'] | null; // Wizard
     wizardNavigate: NexusGenRootTypes['Wizard'] | null; // Wizard
     wizardNavigateForward: NexusGenRootTypes['Wizard'] | null; // Wizard
@@ -188,6 +189,7 @@ export interface NexusGenFieldTypeNames {
     appCreateConfigFile: 'App'
     authenticate: 'App'
     initializePlugins: 'Project'
+    logout: 'App'
     wizardInstallDependencies: 'Wizard'
     wizardNavigate: 'Wizard'
     wizardNavigateForward: 'Wizard'
